Add files via upload
This project involves generating the Mario Kart theme song and converting it into a notes.txt file compatible with a LEGO robot, enabling it to play the tune. The process includes MIDI-to-note conversion and note sequencing tailored to the robot's hardware constraints. Developed as part of a university project in the C++ language .
The file 'LegoSongPlayer.sln' is for generating the sound of the song note by note using an external Library mido for handling midi files and the file 'midi_to_notes.sln' is responsible for generating the txt file that has all the notes with their pitches and velocities and delay time written out . the executable file 'play_notes' is for the ev3 brainstorm environment to play the song with the already installed beep from the robot .