Add files via upload
In this Python program i managed a way that calculates the number of days remaining until a user-specified future date. It prompts the user to enter a date in the "YYYY-MM-DD" format and ensures the input is valid. If the date is correctly formatted and in the future, the program computes the difference between today’s date and the entered date, displaying the number of days left. The program also includes error handling to prevent crashes if the user enters an incorrect format or a past date.